+++ src/time/enums.cpp
@@ -1,4 +1,6 @@
-#include "time_enums.h"
+#include "oa/time/enums.h"
+
+#include <ostream>
 
 #include "helpers/utils.h"
 
@@ -6,11 +8,11 @@ namespace oa::time
 {
     std::ostream& operator<<(std::ostream& os, DayCountRule val)
 	{
-		return os << oa::utils::GetCleanName(val);
+		return os << utils::GetCleanName(val);
 	}
 
 	std::ostream& operator<<(std::ostream& os, AdjRule val)
 	{
-		return os << oa::utils::GetCleanName(val);
+		return os << utils::GetCleanName(val);
 	}
 }
