+++ src/time/time_enums.h
@@ -1,73 +1,14 @@
-#ifndef ODINANALYTICS_TIME_TIMEENUMS_H
-#define ODINANALYTICS_TIME_TIMEENUMS_H
+/**
+ * @file time/time_enums.h
+ * @author MIchael Jang, Derek Huang
+ * @brief C++ header for time enums (compatibility)
+ * @copyright MIT License
+ */
 
-#include "oa/dllexport.h"
+#ifndef OA_TIME_TIME_ENUMS_H_
+#define OA_TIME_TIME_ENUMS_H_
 
-#include <iostream>
+// TODO: refactor existing code using this header to use oa/time/enums.h
+#include "oa/time/enums.h"
 
-namespace oa::time
-{
-	enum class Weekdays
-	{
-		kMonday,
-		kTuesday,
-		kWednesday,
-		kThursday,
-		kFriday,
-		kSaturday,
-		kSunday,
-	};
-
-	enum class Months
-	{
-		kJanuary = 1,
-		kFebruary = 2,
-		kMarch = 3, 
-		kApril = 4, 
-		kMay = 5,
-		kJune = 6, 
-		kJuly = 7,
-		kAugust = 8,
-		kSeptember = 9, 
-		kOctober = 10, 
-		kNovember = 11,
-		kDecember = 12
-	};
-
-	enum class Tenors
-	{
-		kDays = 'D',
-		kWeeks = 'W',
-		kMonths = 'M',
-		kYears = 'Y'
-	};
-
-	enum class AdjRule
-	{
-		kFollowing,
-		kPreceding,
-		kModifiedFollowing,
-		kPlainAdjustment
-	};
-
-	enum class DayCountRule
-	{
-		kACT_ACT,
-		kACTB_ACTB,
-		k30_360_BOND_BASIS,
-		k30_E_360_ISDA,
-		k30_360_E_EUROBOND,
-		kACT_360,
-		kACT_365_FIXED
-
-	};
-	
-    OA_TIME_API
-    std::ostream& operator<<(std::ostream& os, DayCountRule val);
-
-    OA_TIME_API
-    std::ostream& operator<<(std::ostream& os, AdjRule val);
-}
-
-#endif // ODINANALYTICS_TIME_TIMEENUMS_H
-
+#endif  // OA_TIME_TIME_ENUMS_H_
